% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{ALLELES}
\alias{ALLELES}
\title{Germline Allele Names}
\format{An object of class \code{list} of length 2.}
\usage{
ALLELES
}
\description{
Germline allele names that produced the antibodies of interest.
}
\keyword{datasets}
