% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{ANNOTATIONS}
\alias{ANNOTATIONS}
\title{Antibody Annotations}
\format{An object of class \code{list} of length 2.}
\usage{
ANNOTATIONS
}
\description{
Sets of named position ranges within one antibody sequence, currently just
for the heavy chain.
}
\keyword{datasets}
