% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{CDR3}
\alias{CDR3}
\title{CDR[HL]3 amino acid sequences}
\format{An object of class \code{character} of length 2.}
\usage{
CDR3
}
\description{
Amino acid sequences for the CDR3 region of each chain's mature sequences.
Here we've picked a representative sequence in cases where the mature
antibody sequences aren't identical.
}
\keyword{datasets}
