% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{GENES}
\alias{GENES}
\title{Germline Gene Names}
\format{An object of class \code{list} of length 2.}
\usage{
GENES
}
\description{
Gene names for each of the expected germline allele names.
}
\keyword{datasets}
