% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{ggplot_id_div}
\alias{ggplot_id_div}
\title{Plot Identity/Divergence}
\usage{
ggplot_id_div(al, mature, germline, bins = 60)
}
\arguments{
\item{al}{named character vector of already-aligned sequences}

\item{mature}{name of mature reference within al}

\item{germline}{name of germline reference within al}

\item{bins}{number of bins for density plot in either dimension}
}
\value{
ggplot2 plot object
}
\description{
As in SONAR's id/div plots, comparing each sequence to divergence from a
germline sequence (x-axis) and identity to a mature reference (y-axis).
}
