% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{GOOGLE_SHEETS}
\alias{GOOGLE_SHEETS}
\title{Google Sheets IDs}
\format{An object of class \code{list} of length 1.}
\usage{
GOOGLE_SHEETS
}
\description{
A list of workbook names and associated sheet data to track.  Each contained
list should have one entry named "slug" set to the long alphanumeric string
for the workbook, and a second entry named "sheets" containing the integer
GID for each sheet to download with names set to the names of the sheets.
}
\keyword{datasets}
