% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_fasta}
\alias{load_fasta}
\title{Load FASTA file}
\usage{
load_fasta(fp)
}
\arguments{
\item{fp}{FASTA file path}
}
\value{
named character vector
}
\description{
Load a FASTA file into a named character vector
}
