% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_primers}
\alias{load_primers}
\title{Load primer sequence metadata}
\usage{
load_primers(fp = NULL)
}
\arguments{
\item{fp}{file path to sample attributes CSV (defaults to the file built into
the package).}
}
\value{
data frame of primer names and sequences
}
\description{
Load primer sequence metadata
}
