% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_s_attrs}
\alias{load_s_attrs}
\title{Load sample attributes}
\usage{
load_s_attrs(primers, fp = NULL)
}
\arguments{
\item{primers}{data frame of primer metadata to include}

\item{fp}{file path to sample attributes CSV (defaults to the file built into
the package)}
}
\value{
data frame of sample attributes include primers used
}
\description{
Load sample attributes
}
