% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln.R
\name{make_aln_df}
\alias{make_aln_df}
\title{Make data frame version of alignment}
\usage{
make_aln_df(aln_vec, ref = 1, squeeze = NULL)
}
\arguments{
\item{aln_vec}{named character vector of already-aligned sequences}

\item{squeeze}{vector of one or more characters (.e.g, "-").  If any column
contains only a single one of these characters that column will be removed.}

\item{index}{of reference within aln_vec}
}
\description{
Take character vector of already-aligned sequences and create a long-format
data frame with a row for each sequence and position, noting characteristics
relative to a reference sequence within the alignment (by default the first).
}
