% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln.R
\name{make_aln_mat}
\alias{make_aln_mat}
\title{Make matrix version of alignment}
\usage{
make_aln_mat(aln_vec)
}
\arguments{
\item{aln_vec}{character vector of aligned sequences}
}
\value{
matrix of individual characters
}
\description{
Take a sequence alignment in character vector form and return a matrix
equivalent.  If the string lengths aren't identical, shorter sequences will
be padded with NA in the matrix.
}
