% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\name{parse_gene}
\alias{parse_gene}
\title{Parse gene names from allele names}
\usage{
parse_gene(txt)
}
\arguments{
\item{txt}{character vector of immunoglobulin allele names}
}
\value{
character vector of gene names
}
\description{
Parse gene names from allele names
}
