% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{PATHS_REFERENCES}
\alias{PATHS_REFERENCES}
\title{File paths for Reference Sequences}
\format{An object of class \code{list} of length 2.}
\usage{
PATHS_REFERENCES
}
\description{
See REFERENCES for more information.
}
\keyword{datasets}
