% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln.R
\name{plot_aln}
\alias{plot_aln}
\title{Plot Aligned Sequences}
\usage{
plot_aln(aln_vec, ref = 1, aln_df = NULL, draw_text = TRUE,
  style = "vsref")
}
\arguments{
\item{aln_vec}{character vector of already-aligned sequences}

\item{ref}{index of reference within aln_vec}

\item{aln_df}{insted of aln_vec, use this already-prepared data frame}

\item{draw_text}{write text labels on plot?}

\item{style}{keyword for plotting style to use.  One of: dnaplotr, vsref}
}
\value{
ggplot2 plot object
}
\description{
Plot Aligned Sequences
}
