% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{REFERENCES}
\alias{REFERENCES}
\title{Reference Sequences}
\format{An object of class \code{list} of length 2.}
\usage{
REFERENCES
}
\description{
Small sets of sequences for each chain for easy reference.  Each is a named
character vector, stored within a nested list and grouped by chain.
}
\keyword{datasets}
