% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{revcmp}
\alias{revcmp}
\title{Reverse Complement DNA}
\usage{
revcmp(dna)
}
\arguments{
\item{dna}{character vector of DNA nucelotide sequences}
}
\value{
character vector of reverse-complemented sequences
}
\description{
Load a FASTA file into a named character vector
}
