% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{RUNS}
\alias{RUNS}
\title{MiSeq Runs}
\format{An object of class \code{list} of length 3.}
\usage{
RUNS
}
\description{
MiSeq run IDs for the raw data.  At first they're split by heavy/light, but
in the latest set they're mixed.  I'll have to see if this breaks any of my
existing assumptions.
}
\keyword{datasets}
