% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\docType{data}
\name{SAMPLES}
\alias{SAMPLES}
\title{Samples}
\format{An object of class \code{list} of length 2.}
\usage{
SAMPLES
}
\description{
Sample (timepoint) identifiers for heavy and light chains
}
\keyword{datasets}
