% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_csv}
\alias{save_csv}
\title{Save to CSV}
\usage{
save_csv(data, fp)
}
\arguments{
\item{data}{data frame to save}

\item{fp}{file path to save to

Write a data frame to a CSV file with some common settings.}
}
\description{
Save to CSV
}
