% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_fasta}
\alias{save_fasta}
\title{Save FASTA file}
\usage{
save_fasta(txt, fp)
}
\arguments{
\item{txt}{named character vector.  names will be used for sequence
description lines.}

\item{fp}{FASTA file path}
}
\value{
named character vector
}
\description{
Save a FASTA file from a named character vector
}
