% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{sort_seqs}
\alias{sort_seqs}
\title{Sort sequences}
\usage{
sort_seqs(txt, ref = 1, ...)
}
\arguments{
\item{txt}{character vector of aligned sequences}

\item{...}{additional arguments for \code{\link[levenR]{leven}}}
}
\value{
sorted alignment, with earlier sequences more similar to the first
  sequence
}
\description{
Sort a charcter vector of sequences by Levenshtein distance to the first
sequence.
}
