% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{tree_gather_tips}
\alias{tree_gather_tips}
\title{Get tip indexes for a given node}
\usage{
tree_gather_tips(edges, node)
}
\value{
numeric vector of all tip index values linked from a particular node
  index
}
\description{
Get tip indexes for a given node
}
