% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{tree_get_clades}
\alias{tree_get_clades}
\title{Get tips under each node}
\usage{
tree_get_clades(tree)
}
\arguments{
\item{tree}{phylogenetic tree object from \code{\link[ape]{read.tree}}.}
}
\value{
list of character vectors, one for each node (including tips themselves)
}
\description{
Get a vector of tree tip labels for the clade associated with each tree node,
with all the vectors in a list.  The lower half of the list is for the tips
themselves, with the upper half the internal nodes in the tree.  This way the
positions in the list match up with that in the edge matrix for the tree.
}
