% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_igseq.R
\name{update_metadata}
\alias{update_metadata}
\title{Fetch metadata from Google Sheets}
\usage{
update_metadata(google_sheets = GOOGLE_SHEETS, prefix = NULL)
}
\arguments{
\item{google_sheets}{nested list structure like in
\code{\link{GOOGLE_SHEETS}}}

\item{prefix}{directory to put CSV files in.  If \code{NULL}, defaults to the
package's inst directory.}
}
\description{
Fetch metadata from Google Sheets
}
